add env variable irida_url_path to ensure correct war file name (#34)
* add env variable irida_url_path to ensure correct war file name

* check to see if variable set

* better english

+++ tasks/upgrade.sh
@@ -4,6 +4,16 @@ set -e
 
 . /etc/irida/irida_upgrade.config
 
+if [[ -z "${tomcat_location}" ]]; then
+  echo "Value for tomcat_location is not set. Please ensure the value is set in /etc/irida/irida_upgrade.config."
+  exit 1
+fi
+
+if [[ -z "${irida_url_path}" ]]; then
+  echo "Value for irida_url_path is not set. Please ensure the value is set in /etc/irida/irida_upgrade.config."
+  exit 1
+fi
+
 echo "Stopping services..."
 systemctl stop puppet
 systemctl stop tomcat
@@ -19,8 +29,11 @@ chown $tomcat_user:$tomcat_group /tmp/irida.dbbackup
 sudo -u $tomcat_user mv /tmp/irida.dbbackup $db_backup_location/irida-$(date  +"%y-%m-%d-%T").dbbackup
 
 echo "Replacing WAR file..."
-sudo -u $tomcat_user rm $tomcat_location/webapps/irida.war
-sudo -u $tomcat_user curl -o $tomcat_location/webapps/irida.war $war_url
+#removing both expanded folder and war file itself
+sudo -u $tomcat_user rm -rf $tomcat_location/webapps/$irida_url_path
+sudo -u $tomcat_user rm $tomcat_location/webapps/$irida_url_path'.war'
+#downloading newer version
+sudo -u $tomcat_user curl -L -o $tomcat_location/webapps/$irida_url_path'.war' $war_url
 
 echo "Starting services..."
 #only restarting puppet since it will start tomcat for us
